+++ classes/wunderbyte_table.php
@@ -1675,6 +1675,37 @@ class wunderbyte_table extends table_sql {
                     }
                 }
 
+                // Create a list of allowed keys that can be used for filtering.
+                $availablefilters = json_decode($this->filterjson);
+                if (!empty($availablefilters->categories)) {
+                    foreach ($availablefilters->categories as $category) {
+                        $allowedfilters[] = $category->columnname;
+                        // There is an exception for the datepicker: we need to add both the start and end columns as well.
+                        if (
+                            $category->wbfilterclass === 'local_wunderbyte_table\filters\types\datepicker'
+                            &&
+                            is_object($category->datepicker)
+                        ) {
+                            foreach ($category->datepicker->datepickers as $dp) {
+                                if (!empty($dp->startcolumn)) {
+                                    $allowedfilters[] = $dp->startcolumn;
+                                }
+                                if (!empty($dp->endcolumn)) {
+                                    $allowedfilters[] = $dp->endcolumn;
+                                }
+                            }
+                        }
+                    }
+                } else {
+                    $allowedfilters = [];
+                }
+
+                // It’s not important to collect a list of columns that have filters, but rather a list of valid column names
+                // that allows us to compare them with the columns posted by the user for filtering, to prevent injection.
+                // For this reason, we merge any column name from any property that likely holds a column name.
+                $allowedfilters = array_merge($allowedfilters, array_keys($this->columns));
+                $allowedfilters = array_unique($allowedfilters);
+
                 foreach ($categoryvalue as $key => $value) {
                     $filter .= ($categorycounter == 1) ? "" : " AND ";
                     $valuecounter = 1;
@@ -1684,8 +1715,8 @@ class wunderbyte_table extends table_sql {
                             $filter .= ($valuecounter == 1) ? "" : " AND ";
 
                             // In order to make sure we are dealing with real column names and no sql injection...
-                            if (!array_key_exists($categorykey, $this->columns)) {
-                                continue; // Or throw moodle_exception.
+                            if (!in_array($categorykey, $allowedfilters)) {
+                                continue;
                             }
 
                             // We check against allowed operators.
@@ -1693,7 +1724,7 @@ class wunderbyte_table extends table_sql {
                                 'in', 'not in', 'between', 'not between', 'is', 'is not', 'rlike', 'not rlike',
                                 'regexp', 'not regexp', 'ilike', 'not ilike'];
                             if (!in_array($operator, $allowedops, true)) {
-                                continue; // Or throw moodle_exception.
+                                continue;
                             }
 
                             $paramkey = $this->set_params((string)$timestamp, false);